Removed space after last result, fixed mod, fixed overflow

+++ kozyrin.semen/S1/computestr.h
@@ -2,6 +2,6 @@
 #define COMPUTESTR_H
 #include "notation.h"
 
-int computeString(std::string& str);
+long long computeString(std::string& str);
 
 #endif
